+++ firmware/targets/f6/furi-hal/furi-hal-vcp.c
@@ -1,6 +1,7 @@
-#include <furi-hal-vcp.h>
-#include <usbd_cdc_if.h>
+#include <furi-hal-vcp_i.h>
+
 #include <furi.h>
+#include <usbd_cdc_if.h>
 #include <stream_buffer.h>
 
 #define FURI_HAL_VCP_RX_BUFFER_SIZE (APP_RX_DATA_SIZE * 5)
@@ -8,7 +9,7 @@
 extern USBD_HandleTypeDef hUsbDeviceFS;
 
 typedef struct {
-    volatile bool alive;
+    volatile bool connected;
 
     StreamBufferHandle_t rx_stream;
     volatile bool rx_stream_full;
@@ -21,15 +22,9 @@ static FuriHalVcp* furi_hal_vcp = NULL;
 static const uint8_t ascii_soh = 0x01;
 static const uint8_t ascii_eot = 0x04;
 
-void _furi_hal_vcp_init();
-void _furi_hal_vcp_deinit();
-void _furi_hal_vcp_control_line(uint8_t state);
-void _furi_hal_vcp_rx_callback(const uint8_t* buffer, size_t size);
-void _furi_hal_vcp_tx_complete(size_t size);
-
 void furi_hal_vcp_init() {
     furi_hal_vcp = furi_alloc(sizeof(FuriHalVcp));
-    furi_hal_vcp->alive = false;
+    furi_hal_vcp->connected = false;
     
     furi_hal_vcp->rx_stream = xStreamBufferCreate(FURI_HAL_VCP_RX_BUFFER_SIZE, 1);
     furi_hal_vcp->rx_stream_full = false;
@@ -37,53 +32,6 @@ void furi_hal_vcp_init() {
     furi_hal_vcp->tx_semaphore = osSemaphoreNew(1, 1, NULL);
 
     FURI_LOG_I("FuriHalVcp", "Init OK");
-}
-
-void _furi_hal_vcp_init() {
-    osSemaphoreRelease(furi_hal_vcp->tx_semaphore);
-}
-
-void _furi_hal_vcp_deinit() {
-    furi_hal_vcp->alive = false;
-    osSemaphoreRelease(furi_hal_vcp->tx_semaphore);
-}
-
-void _furi_hal_vcp_control_line(uint8_t state) {
-    // bit 0: DTR state, bit 1: RTS state
-    // bool dtr = state & 0b01;
-    bool dtr = state & 0b1;
-
-    if (dtr) {
-        if (!furi_hal_vcp->alive) {
-            furi_hal_vcp->alive = true;
-            _furi_hal_vcp_rx_callback(&ascii_soh, 1); // SOH
-        }
-    } else {
-        if (furi_hal_vcp->alive) {
-            _furi_hal_vcp_rx_callback(&ascii_eot, 1); // EOT
-            furi_hal_vcp->alive = false;
-        }
-    }
-
-    osSemaphoreRelease(furi_hal_vcp->tx_semaphore);
-}
-
-void _furi_hal_vcp_rx_callback(const uint8_t* buffer, size_t size) {
-    BaseType_t xHigherPriorityTaskWoken = pdFALSE;
-    size_t ret = xStreamBufferSendFromISR(furi_hal_vcp->rx_stream, buffer, size, &xHigherPriorityTaskWoken);
-    furi_check(ret == size);
-    
-    if (xStreamBufferSpacesAvailable(furi_hal_vcp->rx_stream) >= APP_RX_DATA_SIZE) {
-        USBD_CDC_ReceivePacket(&hUsbDeviceFS);
-    } else {
-        furi_hal_vcp->rx_stream_full = true;
-    }
-
-    portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
-}
-
-void _furi_hal_vcp_tx_complete(size_t size) {
-    osSemaphoreRelease(furi_hal_vcp->tx_semaphore);
 }
 
 size_t furi_hal_vcp_rx(uint8_t* buffer, size_t size) {
@@ -109,8 +57,10 @@ size_t furi_hal_vcp_rx_with_timeout(uint8_t* buffer, size_t size, uint32_t timeo
 void furi_hal_vcp_tx(const uint8_t* buffer, size_t size) {
     furi_assert(furi_hal_vcp);
 
-    while (size > 0 && furi_hal_vcp->alive) {
+    while (size > 0 && furi_hal_vcp->connected) {
         furi_check(osSemaphoreAcquire(furi_hal_vcp->tx_semaphore, osWaitForever) == osOK);
+        if (!furi_hal_vcp->connected)
+            break;
 
         size_t batch_size = size;
         if (batch_size > APP_TX_DATA_SIZE) {
@@ -121,8 +71,58 @@ void furi_hal_vcp_tx(const uint8_t* buffer, size_t size) {
             size -= batch_size;
             buffer += batch_size;
         } else {
-            // Shouldn't be there 
-            osDelay(100);
+            FURI_LOG_E("FuriHalVcp", "CDC_Transmit_FS failed");
+            osDelay(50);
         }
     }
 }
+
+void furi_hal_vcp_on_usb_resume() {
+    osSemaphoreRelease(furi_hal_vcp->tx_semaphore);
+}
+
+void furi_hal_vcp_on_usb_suspend() {
+    if (furi_hal_vcp->connected) {
+        furi_hal_vcp->connected = false;
+        osSemaphoreRelease(furi_hal_vcp->tx_semaphore);
+    }
+}
+
+void furi_hal_vcp_on_cdc_control_line(uint8_t state) {
+    // bit 0: DTR state, bit 1: RTS state
+    // bool dtr = state & 0b01;
+    bool dtr = state & 0b1;
+
+    if (dtr) {
+        if (!furi_hal_vcp->connected) {
+            furi_hal_vcp->connected = true;
+            furi_hal_vcp_on_cdc_rx(&ascii_soh, 1); // SOH
+        }
+    } else {
+        if (furi_hal_vcp->connected) {
+            furi_hal_vcp_on_cdc_rx(&ascii_eot, 1); // EOT
+            furi_hal_vcp->connected = false;
+        }
+    }
+
+    osSemaphoreRelease(furi_hal_vcp->tx_semaphore);
+}
+
+void furi_hal_vcp_on_cdc_rx(const uint8_t* buffer, size_t size) {
+    BaseType_t xHigherPriorityTaskWoken = pdFALSE;
+    size_t ret = xStreamBufferSendFromISR(furi_hal_vcp->rx_stream, buffer, size, &xHigherPriorityTaskWoken);
+    furi_check(ret == size);
+    
+    if (xStreamBufferSpacesAvailable(furi_hal_vcp->rx_stream) >= APP_RX_DATA_SIZE) {
+        USBD_CDC_ReceivePacket(&hUsbDeviceFS);
+    } else {
+        furi_hal_vcp->rx_stream_full = true;
+    }
+
+    portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
+}
+
+void furi_hal_vcp_on_cdc_tx_complete(size_t size) {
+    osSemaphoreRelease(furi_hal_vcp->tx_semaphore);
+}
+
