+++ firmware/targets/f6/usb-glue/usb_device.h
