+++ firmware/targets/f6/usb-glue/usbd_conf.c
@@ -1,0 +1,506 @@
+#include "stm32wbxx.h"
+#include "stm32wbxx_hal.h"
+
+#include <furi-hal-vcp_i.h>
+
+#include "usbd_def.h"
+#include "usbd_core.h"
+#include "usbd_cdc.h"
+
+PCD_HandleTypeDef hpcd_USB_FS;
+void Error_Handler(void);
+
+static USBD_StatusTypeDef USBD_Get_USB_Status(HAL_StatusTypeDef hal_status);
+
+static void SystemClockConfig_Resume(void);
+
+void HAL_PCD_MspInit(PCD_HandleTypeDef* pcdHandle) {
+    GPIO_InitTypeDef GPIO_InitStruct = {0};
+    if(pcdHandle->Instance==USB) {
+        __HAL_RCC_GPIOA_CLK_ENABLE();
+        /**USB GPIO Configuration
+        PA11     ------> USB_DM
+        PA12     ------> USB_DP 
+        */
+        GPIO_InitStruct.Pin = GPIO_PIN_11|GPIO_PIN_12;
+        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
+        GPIO_InitStruct.Pull = GPIO_NOPULL;
+        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
+        GPIO_InitStruct.Alternate = GPIO_AF10_USB;
+        HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
+
+        /* Peripheral clock enable */
+        __HAL_RCC_USB_CLK_ENABLE();
+
+        /* Peripheral interrupt init */
+        HAL_NVIC_SetPriority(USB_LP_IRQn, 5, 0);
+        HAL_NVIC_EnableIRQ(USB_LP_IRQn);
+    }
+}
+
+void HAL_PCD_MspDeInit(PCD_HandleTypeDef* pcdHandle) {
+    if(pcdHandle->Instance==USB) {
+        /* Peripheral clock disable */
+        __HAL_RCC_USB_CLK_DISABLE();
+
+        /**USB GPIO Configuration
+        PA11     ------> USB_DM
+        PA12     ------> USB_DP
+        */
+        HAL_GPIO_DeInit(GPIOA, GPIO_PIN_11|GPIO_PIN_12);
+
+        /* Peripheral interrupt Deinit*/
+        HAL_NVIC_DisableIRQ(USB_LP_IRQn);
+    }
+}
+
+/** Setup stage callback
+ * @param  hpcd: PCD handle
+ * @retval None
+ */
+void HAL_PCD_SetupStageCallback(PCD_HandleTypeDef *hpcd) {
+    USBD_LL_SetupStage((USBD_HandleTypeDef*)hpcd->pData, (uint8_t *)hpcd->Setup);
+}
+
+/** Data Out stage callback.
+ * @param  hpcd: PCD handle
+ * @param  epnum: Endpoint number
+ * @retval None
+ */
+void HAL_PCD_DataOutStageCallback(PCD_HandleTypeDef *hpcd, uint8_t epnum) {
+    USBD_LL_DataOutStage((USBD_HandleTypeDef*)hpcd->pData, epnum, hpcd->OUT_ep[epnum].xfer_buff);
+}
+
+/** Data In stage callback.
+ * @param  hpcd: PCD handle
+ * @param  epnum: Endpoint number
+ * @retval None
+ */
+void HAL_PCD_DataInStageCallback(PCD_HandleTypeDef *hpcd, uint8_t epnum) {
+    USBD_LL_DataInStage((USBD_HandleTypeDef*)hpcd->pData, epnum, hpcd->IN_ep[epnum].xfer_buff);
+}
+
+/** SOF callback.
+ * @param  hpcd: PCD handle
+ * @retval None
+ */
+void HAL_PCD_SOFCallback(PCD_HandleTypeDef *hpcd) {
+    USBD_LL_SOF((USBD_HandleTypeDef*)hpcd->pData);
+}
+
+/** Reset callback.
+ * @param  hpcd: PCD handle
+ * @retval None
+ */
+void HAL_PCD_ResetCallback(PCD_HandleTypeDef *hpcd) {
+    USBD_SpeedTypeDef speed = USBD_SPEED_FULL;
+
+    if ( hpcd->Init.speed != PCD_SPEED_FULL) {
+        Error_Handler();
+    }
+
+    /* Set Speed. */
+    USBD_LL_SetSpeed((USBD_HandleTypeDef*)hpcd->pData, speed);
+
+    /* Reset Device. */
+    USBD_LL_Reset((USBD_HandleTypeDef*)hpcd->pData);
+}
+
+/** Suspend callback.
+ * When Low power mode is enabled the debug cannot be used (IAR, Keil doesn't support it)
+ * @param  hpcd: PCD handle
+ * @retval None
+ */
+void HAL_PCD_SuspendCallback(PCD_HandleTypeDef *hpcd) {
+    USBD_LL_Suspend((USBD_HandleTypeDef*)hpcd->pData);
+
+    furi_hal_vcp_on_usb_suspend();
+    
+    if (hpcd->Init.low_power_enable) {
+        /* Set SLEEPDEEP bit and SleepOnExit of Cortex System Control Register. */
+        SCB->SCR |= (uint32_t)((uint32_t)(SCB_SCR_SLEEPDEEP_Msk | SCB_SCR_SLEEPONEXIT_Msk));
+    }
+}
+
+/** Resume callback.
+ * When Low power mode is enabled the debug cannot be used (IAR, Keil doesn't support it)
+ * @param  hpcd: PCD handle
+ * @retval None
+ */
+void HAL_PCD_ResumeCallback(PCD_HandleTypeDef *hpcd) {
+    if (hpcd->Init.low_power_enable) {
+        /* Reset SLEEPDEEP bit of Cortex System Control Register. */
+        SCB->SCR &= (uint32_t)~((uint32_t)(SCB_SCR_SLEEPDEEP_Msk | SCB_SCR_SLEEPONEXIT_Msk));
+        SystemClockConfig_Resume();
+    }
+
+    furi_hal_vcp_on_usb_resume();
+
+    USBD_LL_Resume((USBD_HandleTypeDef*)hpcd->pData);
+}
+
+/** ISOOUTIncomplete callback.
+ * @param  hpcd: PCD handle
+ * @param  epnum: Endpoint number
+ * @retval None
+ */
+void HAL_PCD_ISOOUTIncompleteCallback(PCD_HandleTypeDef *hpcd, uint8_t epnum) {
+    USBD_LL_IsoOUTIncomplete((USBD_HandleTypeDef*)hpcd->pData, epnum);
+}
+
+/** ISOINIncomplete callback.
+ * @param  hpcd: PCD handle
+ * @param  epnum: Endpoint number
+ * @retval None
+ */
+void HAL_PCD_ISOINIncompleteCallback(PCD_HandleTypeDef *hpcd, uint8_t epnum) {
+    USBD_LL_IsoINIncomplete((USBD_HandleTypeDef*)hpcd->pData, epnum);
+}
+
+/** Connect callback.
+ * @param  hpcd: PCD handle
+ * @retval None
+ */
+void HAL_PCD_ConnectCallback(PCD_HandleTypeDef *hpcd) {
+    USBD_LL_DevConnected((USBD_HandleTypeDef*)hpcd->pData);
+}
+
+/** Disconnect callback.
+ * @param  hpcd: PCD handle
+ * @retval None
+ */
+void HAL_PCD_DisconnectCallback(PCD_HandleTypeDef *hpcd) {
+    USBD_LL_DevDisconnected((USBD_HandleTypeDef*)hpcd->pData);
+}
+
+/** Initializes the low level portion of the device driver.
+ * @param  pdev: Device handle
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_Init(USBD_HandleTypeDef *pdev) {
+    /* Init USB Ip. */
+    hpcd_USB_FS.pData = pdev;
+
+    /* Link the driver to the stack. */
+    pdev->pData = &hpcd_USB_FS;
+
+    /* Enable USB power on Pwrctrl CR2 register. */
+    HAL_PWREx_EnableVddUSB();
+
+    hpcd_USB_FS.Instance = USB;
+    hpcd_USB_FS.Init.dev_endpoints = 8;
+    hpcd_USB_FS.Init.speed = PCD_SPEED_FULL;
+    hpcd_USB_FS.Init.phy_itface = PCD_PHY_EMBEDDED;
+    hpcd_USB_FS.Init.Sof_enable = DISABLE;
+    hpcd_USB_FS.Init.low_power_enable = DISABLE;
+    hpcd_USB_FS.Init.lpm_enable = DISABLE;
+    hpcd_USB_FS.Init.battery_charging_enable = DISABLE;
+
+    if (HAL_PCD_Init(&hpcd_USB_FS) != HAL_OK) {
+        Error_Handler( );
+    }
+
+    HAL_PCDEx_PMAConfig((PCD_HandleTypeDef*)pdev->pData , 0x00 , PCD_SNG_BUF, 0x18);
+    HAL_PCDEx_PMAConfig((PCD_HandleTypeDef*)pdev->pData , 0x80 , PCD_SNG_BUF, 0x58);
+
+    HAL_PCDEx_PMAConfig((PCD_HandleTypeDef*)pdev->pData , 0x81 , PCD_SNG_BUF, 0xC0);
+    HAL_PCDEx_PMAConfig((PCD_HandleTypeDef*)pdev->pData , 0x01 , PCD_SNG_BUF, 0x110);
+    HAL_PCDEx_PMAConfig((PCD_HandleTypeDef*)pdev->pData , 0x82 , PCD_SNG_BUF, 0x100);
+
+    return USBD_OK;
+}
+
+/** De-Initializes the low level portion of the device driver.
+ * @param  pdev: Device handle
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_DeInit(USBD_HandleTypeDef *pdev)
+{
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_DeInit(pdev->pData);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Starts the low level portion of the device driver.
+ * @param  pdev: Device handle
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_Start(USBD_HandleTypeDef *pdev) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_Start(pdev->pData);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Stops the low level portion of the device driver.
+ * @param  pdev: Device handle
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_Stop(USBD_HandleTypeDef *pdev) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_Stop(pdev->pData);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Opens an endpoint of the low level driver.
+ * @param  pdev: Device handle
+ * @param  ep_addr: Endpoint number
+ * @param  ep_type: Endpoint type
+ * @param  ep_mps: Endpoint max packet size
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_OpenEP(USBD_HandleTypeDef *pdev, uint8_t ep_addr, uint8_t ep_type, uint16_t ep_mps) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_EP_Open(pdev->pData, ep_addr, ep_mps, ep_type);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Closes an endpoint of the low level driver.
+ * @param  pdev: Device handle
+ * @param  ep_addr: Endpoint number
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_CloseEP(USBD_HandleTypeDef *pdev, uint8_t ep_addr) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_EP_Close(pdev->pData, ep_addr);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/**
+ * @brief  Flushes an endpoint of the Low Level Driver.
+ * @param  pdev: Device handle
+ * @param  ep_addr: Endpoint number
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_FlushEP(USBD_HandleTypeDef *pdev, uint8_t ep_addr) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_EP_Flush(pdev->pData, ep_addr);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Sets a Stall condition on an endpoint of the Low Level Driver.
+ * @param  pdev: Device handle
+ * @param  ep_addr: Endpoint number
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_StallEP(USBD_HandleTypeDef *pdev, uint8_t ep_addr) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_EP_SetStall(pdev->pData, ep_addr);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Clears a Stall condition on an endpoint of the Low Level Driver.
+ * @param  pdev: Device handle
+ * @param  ep_addr: Endpoint number
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_ClearStallEP(USBD_HandleTypeDef *pdev, uint8_t ep_addr) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_EP_ClrStall(pdev->pData, ep_addr);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Returns Stall condition.
+ * @param  pdev: Device handle
+ * @param  ep_addr: Endpoint number
+ * @retval Stall (1: Yes, 0: No)
+ */
+uint8_t USBD_LL_IsStallEP(USBD_HandleTypeDef *pdev, uint8_t ep_addr) {
+    PCD_HandleTypeDef *hpcd = (PCD_HandleTypeDef*) pdev->pData;
+
+    if((ep_addr & 0x80) == 0x80)
+    {
+        return hpcd->IN_ep[ep_addr & 0x7F].is_stall;
+    }
+    else
+    {
+        return hpcd->OUT_ep[ep_addr & 0x7F].is_stall;
+    }
+}
+
+/** Assigns a USB address to the device.
+ * @param  pdev: Device handle
+ * @param  dev_addr: Device address
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_SetUSBAddress(USBD_HandleTypeDef *pdev, uint8_t dev_addr) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_SetAddress(pdev->pData, dev_addr);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Transmits data over an endpoint.
+ * @param  pdev: Device handle
+ * @param  ep_addr: Endpoint number
+ * @param  pbuf: Pointer to data to be sent
+ * @param  size: Data size
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_Transmit(USBD_HandleTypeDef *pdev, uint8_t ep_addr, uint8_t *pbuf, uint32_t size) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_EP_Transmit(pdev->pData, ep_addr, pbuf, size);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Prepares an endpoint for reception.
+ * @param  pdev: Device handle
+ * @param  ep_addr: Endpoint number
+ * @param  pbuf: Pointer to data to be received
+ * @param  size: Data size
+ * @retval USBD status
+ */
+USBD_StatusTypeDef USBD_LL_PrepareReceive(USBD_HandleTypeDef *pdev, uint8_t ep_addr, uint8_t *pbuf, uint32_t size) {
+    HAL_StatusTypeDef hal_status = HAL_OK;
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    hal_status = HAL_PCD_EP_Receive(pdev->pData, ep_addr, pbuf, size);
+
+    usb_status =  USBD_Get_USB_Status(hal_status);
+
+    return usb_status;
+}
+
+/** Returns the last transfered packet size.
+ * @param  pdev: Device handle
+ * @param  ep_addr: Endpoint number
+ * @retval Recived Data Size
+ */
+uint32_t USBD_LL_GetRxDataSize(USBD_HandleTypeDef *pdev, uint8_t ep_addr) {
+    return HAL_PCD_EP_GetRxCount((PCD_HandleTypeDef*) pdev->pData, ep_addr);
+}
+
+/** Send LPM message to user layer
+ * @param  hpcd: PCD handle
+ * @param  msg: LPM message
+ * @retval None
+ */
+void HAL_PCDEx_LPM_Callback(PCD_HandleTypeDef *hpcd, PCD_LPM_MsgTypeDef msg) {
+    switch (msg) {
+    case PCD_LPM_L0_ACTIVE:
+        if (hpcd->Init.low_power_enable) {
+            SystemClockConfig_Resume();
+            /* Reset SLEEPDEEP bit of Cortex System Control Register. */
+            SCB->SCR &= (uint32_t)~((uint32_t)(SCB_SCR_SLEEPDEEP_Msk | SCB_SCR_SLEEPONEXIT_Msk));
+        }
+        USBD_LL_Resume(hpcd->pData);
+        break;
+
+    case PCD_LPM_L1_ACTIVE:
+        USBD_LL_Suspend(hpcd->pData);
+
+        /* Enter in STOP mode. */
+        if (hpcd->Init.low_power_enable) {
+            /* Set SLEEPDEEP bit and SleepOnExit of Cortex System Control Register. */
+            SCB->SCR |= (uint32_t)((uint32_t)(SCB_SCR_SLEEPDEEP_Msk | SCB_SCR_SLEEPONEXIT_Msk));
+        }
+        break;
+    }
+}
+
+/** Delays routine for the USB Device Library.
+ * @param  Delay: Delay in ms
+ * @retval None
+ */
+void USBD_LL_Delay(uint32_t Delay) {
+    HAL_Delay(Delay);
+}
+
+/** Static single allocation.
+ * @param  size: Size of allocated memory
+ * @retval None
+ */
+void *USBD_static_malloc(uint32_t size) {
+    static uint32_t mem[(sizeof(USBD_CDC_HandleTypeDef)/4)+1];/* On 32-bit boundary */
+    return mem;
+}
+
+/** Dummy memory free
+ * @param  p: Pointer to allocated  memory address
+ * @retval None
+ */
+void USBD_static_free(void *p) {
+}
+
+/** Configures system clock after wake-up from USB resume callBack:
+ * enable HSI, PLL and select PLL as system clock source.
+ * @retval None
+ */
+static void SystemClockConfig_Resume(void) {
+}
+
+/** Retuns the USB status depending on the HAL status:
+ * @param  hal_status: HAL status
+ * @retval USB status
+ */
+USBD_StatusTypeDef USBD_Get_USB_Status(HAL_StatusTypeDef hal_status) {
+    USBD_StatusTypeDef usb_status = USBD_OK;
+
+    switch (hal_status)
+    {
+        case HAL_OK :
+            usb_status = USBD_OK;
+        break;
+        case HAL_ERROR :
+            usb_status = USBD_FAIL;
+        break;
+        case HAL_BUSY :
+            usb_status = USBD_BUSY;
+        break;
+        case HAL_TIMEOUT :
+            usb_status = USBD_FAIL;
+        break;
+        default :
+            usb_status = USBD_FAIL;
+        break;
+    }
+    return usb_status;
+}
