+++ firmware/targets/f6/usb-glue/usbd_desc.c
